Abstract Field context logic with a HOC
